+++ tiledb/sm/query/query.h
@@ -494,8 +494,9 @@ class Query {
   Status set_config(const Config& config);
 
   /**
-   * Get the config of the query
-   * @return
+   * Get the config of the query.
+   *
+   * @return Config from query
    */
   const Config& config() const;